:refactor: move parse_time function to private utils and update references
