+++ discord/integrations.py
@@ -31,8 +31,8 @@ from typing import TYPE_CHECKING, Any
 from .enums import ExpireBehaviour, try_enum
 from .errors import InvalidArgument
 from .user import User
-from .utils import MISSING, parse_time
-from .utils.private import get_as_snowflake
+from .utils import MISSING
+from .utils.private import get_as_snowflake, parse_time
 from discord import utils
 
 __all__ = (
