+++ discord/threads.py
@@ -32,8 +32,8 @@ from .enums import ChannelType, try_enum
 from .errors import ClientException
 from .flags import ChannelFlags
 from .mixins import Hashable
-from .utils import MISSING, parse_time
-from .utils.private import get_as_snowflake
+from .utils import MISSING
+from .utils.private import get_as_snowflake, parse_time
 from discord import utils
 
 __all__ = (
