+++ test.py
@@ -2,15 +2,15 @@ import discord
 import os
 from dotenv import load_dotenv
 
-from discord.ext import commands
-
 load_dotenv()
 
-bot = commands.Bot()
+bot = discord.Bot()
+
 
 @bot.slash_command()
-async def hi(ctx: discord.ApplicationContext, name: str | None = "World"):
-    """Say hi to someone."""
-    await ctx.respond(f"Hello {name}!")
+async def hi(ctx: discord.ApplicationContext):
+    """Test command to show user information."""
+    await ctx.respond(f"Hello {ctx.author.name}!")
 
-bot.run(os.getenv(("TOKEN_2")))
+
+bot.run(os.getenv("TOKEN_2"))
